==> notebooks/00_functions.ipynb <==
import pandas as pd
import numpy as np
from sklearn.cluster import KMeans

df = pd.read_csv(filepath_or_buffer='../data/taxi_zones.csv')
# df = df[['the_geom', 'zone']]
df.columns

def calculate_midpoint(coordinates):
    # Convert coordinates to numpy array
    coordinates = np.array(coordinates)
    
    # Perform spatial clustering using K-means
    kmeans = KMeans(n_clusters=1).fit(coordinates)
    
    # Get the centroid of the cluster
    centroid = kmeans.cluster_centers_[0]
    
    return centroid

def get_coordinates(df: pd.DataFrame):
    df = df[['the_geom', 'zone', 'LocationID']]
    lat_list = []
    long_list = []

    for i in range(0,263):
        l = df['the_geom'][i][16:].replace(',','').replace('(','').replace(')','').split()

        # long = [float(l[j]) for j in range(len(l)) if j % 2 == 0]
        # lat = [float(l[j]) for j in range(len(l)) if j % 2 != 0]
        # mean_lat = np.mean([max(lat),min(lat)])
        # mean_long = np.mean([max(long),min(long)])

        l = [float(x) for x in l]
        l = [l[i:i+2] for i in range(0, len(l), 2)]

        centroid = calculate_midpoint(l)

        lat_list.append(centroid[1])
        long_list.append(centroid[0])

    coords = pd.DataFrame(list(zip(lat_list, long_list)),columns=['Latitude', 'Longitude'])

    return pd.concat([df[['zone','LocationID']], coords],axis=1)

get_coordinates(df)

l = df['the_geom'][0][16:].replace(',','').replace('(','').replace(')','').split()

from sklearn.cluster import KMeans
import numpy as np

def calculate_midpoint(coordinates):
    # Convert coordinates to numpy array
    coordinates = np.array(coordinates)
    
    # Perform spatial clustering using K-means
    kmeans = KMeans(n_clusters=1).fit(coordinates)
    
    # Get the centroid of the cluster
    centroid = kmeans.cluster_centers_[0]
    
    return centroid

# calculate_midpoint(l)

l = [float(x) for x in l]

l = [l[i:i+2] for i in range(0, len(l), 2)]

calculate_midpoint(l)

l = df['the_geom'][32][16:-3].replace(',','').replace('(','').replace(')','').split()

long = [float(l[i]) for i in range(len(l)) if i % 2 == 0]
lat = [float(l[i]) for i in range(len(l)) if i % 2 != 0]

mean_lat = np.mean([max(lat),min(lat)])
mean_long = np.mean([max(long),min(long)])

print(f'{mean_lat}, {mean_long}')

from model import get_pipeline

get_pipeline()
