==> notebooks/01_load_and_validate_raw_data.ipynb <==
import numpy as np
import matplotlib.pyplot as plt
import torch.nn.functional as F
import pandas as pd
from pathlib import Path
import requests

# Preprocessing for the machine learning data

from pathlib import Path
import requests

def download_file(year: int , month: int) -> Path:
    """"""
    URL = f'https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet'
    response = requests.get(URL)

    if response.status_code == 200:
        path = f"../data/raw/rides_{year}-{month:02d}.parquet"
        open(path, 'wb').write(response.content)
        return path
    else:
        raise Exception(f"Failed to download {URL}")

download_file(2022,1)

rides = pd.read_parquet("../data/raw/rides_2022-01.parquet")

# rides.head(25) show the entire rides dataframe

rides = rides[["tpep_pickup_datetime","PULocationID"]]

rides.rename(columns = {
    "tpep_pickup_datetime":"pickup_daytime",
    "PULocationID": "pickup_location_id"
},inplace=True)

rides.head()

# Data validation and data cleaning:
## test for inaccuracies in the data labelling of the dates

rides['pickup_daytime'].describe()

rides = rides[rides.pickup_daytime >= '2022-01-01']
rides = rides[rides.pickup_daytime < '2022-02-01']

rides['pickup_daytime'].describe()

## read the data to a new file of transformed data

rides.to_parquet('../data/transformed/cleaned_rides_2022-01.parquet')
