==> notebooks/02_transform_raw_data_to_time_series.ipynb <==
import pandas as pd
import pyarrow

rides = pd.read_parquet('../data/transformed/cleaned_rides_2022-01.parquet')
rides.head(10) # each row contains the information for one ride

# add column to collect hourly data
rides['pickup_hour'] = rides['pickup_daytime'].dt.floor('H') # dt.round('H') rounds datetime64 data to the closest hour / alternative dt.floor('H') but that did not work for me
rides

# aggregate the rides across the time and location data 

agg_rides = rides.groupby(['pickup_hour','pickup_location_id']).size().reset_index()
agg_rides.rename(columns={0:'rides'}, inplace=True)
agg_rides.sort_values(by=['pickup_hour','pickup_location_id'])

import pandas as pd
from tqdm import tqdm

def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:
    '''
    https://stackoverflow.com/questions/19324453/add-missing-dates-to-pandas-dataframe/19324591#19324591
    '''
    location_ids = agg_rides['pickup_location_id'].unique()
    full_range = pd.date_range(
        agg_rides['pickup_hour'].min(), agg_rides['pickup_hour'].max(), freq='H')
    output = pd.DataFrame()

    for location_id in tqdm(location_ids):

        agg_rides_i = agg_rides.loc[agg_rides.pickup_location_id == location_id, ['pickup_hour','rides']]

        agg_rides_i.set_index('pickup_hour', inplace=True)
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0)

        agg_rides_i['pickup_location_id'] = location_id

        output = pd.concat([output, agg_rides_i])

    output = output.reset_index().rename(columns={"index": 'pickup_hour'})

    return output


agg_rides_all_slots = add_missing_slots(agg_rides)


agg_rides_all_slots

from typing import Optional, List
import plotly.express as px

def plot_rides(
    rides: pd.DataFrame,
    locations: Optional[List[int]] = None
    ):
    """
    Plot rides.

    Parameters
    ----------
    rides : pd.DataFrame
        Rides dataframe.
    locations : Optional[List[int]], optional
        List of locations to plot. The default is None.

    Returns
    -------
    fig : plotly.graph_objects.Figure
        Plotly figure.
    """

    rides_to_plot = rides[rides.pickup_location_id.isin(locations)] if locations else rides

    fig = px.line(
        rides_to_plot,
        x="pickup_hour",
        y="rides",
        color="pickup_location_id",
        template='none'
    )

    fig.show()

    return None

plot_rides(agg_rides_all_slots, [1])

agg_rides_all_slots.to_parquet('../data/transformed/ts_data_2022_01.parquet')
