==> notebooks/04_transform_raw_data_into_features_and_targets.ipynb <==
import pandas as pd
import numpy as np
from functions import *
from pathlib import Path

# autoreload
%reload_ext autoreload
%autoreload 2


#########################################
# Full data pipeline for the ML project #
#########################################

# Step 1: Import the data from an external source; this uses the function in data.py called load_raw_data

rides = load_raw_data(year=2022, months=None)
rides

# Step 2: Create the time series data from the cleaned raw data

ts_data = transform_raw_data_into_ts_data(rides=rides)
print(ts_data)

features, targets = transform_ts_data_into_features_and_target(ts_data=ts_data,
                                                               input_seq_len=24*28*1, # one month
                                                               step_size=24) 

print(f'{features.shape=}')
print(f'{targets.shape=}')

tabular_data = features
tabular_data['target_rides_next_hour'] = targets
tabular_data.to_parquet(CLEANED_DATA_DIR / 'tabular_data.parquet')

pd.set_option('display.max_rows', 10)
tabular_data
