==> notebooks/05_visualize_training_data.ipynb <==
import pandas as pd
from paths import CLEANED_DATA_DIR
from functions import plot_one_sample

# autoreload
%reload_ext autoreload
%autoreload 2

# Tabular Dataframe 

features_and_targets = pd.read_parquet(CLEANED_DATA_DIR / 'tabular_data.parquet')
features_and_targets.head(15)

features = features_and_targets.drop(columns='target_rides_next_hour')
targets = features_and_targets['target_rides_next_hour']

plot_one_sample(5,features,targets)
