==> notebooks/09_catboost_model_with_feature_engineering.ipynb <==
import pandas as pd
import numpy as np
from paths import CLEANED_DATA_DIR
from datetime import datetime
from data_split import train_test_split
import xgboost as xgb
from sklearn.metrics import mean_absolute_error
from sklearn.preprocessing import FunctionTransformer
from sklearn.base import BaseEstimator, TransformerMixin
import catboost

def get_coordinates(df: pd.DataFrame):
    df = df[['the_geom', 'zone', 'LocationID']]
    df = df.rename(columns={'LocationID': 'pickup_location_id'})
    lat_list = []
    long_list = []

    for i in range(0,263):
        l = df['the_geom'][i][16:].replace(',','').replace('(','').replace(')','').split()

        long = [float(l[j]) for j in range(len(l)) if j % 2 == 0]
        lat = [float(l[j]) for j in range(len(l)) if j % 2 != 0]
        mean_lat = np.mean([max(lat),min(lat)])
        mean_long = np.mean([max(long),min(long)])

        lat_list.append(mean_lat)
        long_list.append(mean_long)

    coords = pd.DataFrame(list(zip(lat_list, long_list)),columns=['Latitude', 'Longitude'])

    return pd.concat([df[['pickup_location_id']], coords],axis=1)



coord_df = pd.read_csv(filepath_or_buffer='../data/taxi_zones.csv')
coord_df = get_coordinates(coord_df)

df = pd.read_parquet(CLEANED_DATA_DIR / 'tabular_data.parquet')
df = df.merge(coord_df, on='pickup_location_id', how='left')
df.drop(columns=['pickup_location_id'],inplace=True)

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 11, 1, 0, 0, 0),
    target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

df.head(1)

def average_rides_last_4_weeks(X: pd.DataFrame) -> pd.DataFrame:
    """
    Adds one column with the average rides from
    - 7 days ago
    - 14 days ago
    - 21 days ago
    - 28 days ago
    """
    X['average_rides_last_4_weeks'] = 0.25*(
        X[f'rides_previous_{7*24}_hour'] + \
        X[f'rides_previous_{2*7*24}_hour'] + \
        X[f'rides_previous_{3*7*24}_hour'] + \
        X[f'rides_previous_{4*7*24}_hour']
    )
    return X

add_feature_average_rides_last_4_weeks = FunctionTransformer(
    average_rides_last_4_weeks, validate=False)

add_feature_average_rides_last_4_weeks.fit_transform(X_train)

class TemporalFeaturesEngineer(BaseEstimator, TransformerMixin):

    def fit(self, X, y=None):
        return self
    
    def transform(self, X, y=None):
        
        X_ = X.copy()
        
        # Generate numeric columns from datetime
        X_["hour"] = X_['pickup_hour'].dt.hour
        X_["day_of_week"] = X_['pickup_hour'].dt.dayofweek
        
        return X_.drop(columns=['pickup_hour'])

add_temporal_features = TemporalFeaturesEngineer()
add_temporal_features.fit_transform(X_train)

from sklearn.pipeline import make_pipeline

pipeline = make_pipeline(
    add_feature_average_rides_last_4_weeks,
    add_temporal_features,
    catboost.CatBoostRegressor()
)
pipeline.fit(X_train, y_train)

predictions = pipeline.predict(X_test)

from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')
