==> notebooks/10_catboost_with_hyperparameter_tuning.ipynb <==
import pandas as pd
import numpy as np
from paths import CLEANED_DATA_DIR
from datetime import datetime
from data_split import train_test_split
import xgboost as xgb
from sklearn.metrics import mean_absolute_error
from sklearn.preprocessing import FunctionTransformer
from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.model_selection import KFold, TimeSeriesSplit
from sklearn.pipeline import make_pipeline
from sklearn.metrics import mean_absolute_error
import optuna
import catboost
import xgboost as xgb

from model import *


# autoreload
%reload_ext autoreload
%autoreload 2

df = pd.read_parquet(CLEANED_DATA_DIR / 'tabular_data.parquet')
df = add_coordinates(df)

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 6, 1, 0, 0, 0),
    target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

tss = TimeSeriesSplit(n_splits=2)
scores = []
for train_index, val_index in tss.split(X_train):

    # split data for training and validation
    X_train_, X_val_ = X_train.iloc[train_index, :], X_train.iloc[val_index,:]
    y_train_, y_val_ = y_train.iloc[train_index], y_train.iloc[val_index]

    
    print(f'{X_train_.shape=}')      
    print(f'{y_train_.shape=}')
    print(f'{X_val_.shape=}')
    print(f'{y_val_.shape=}')

def objective(trial: optuna.trial.Trial) -> float:
    """
    Given a set of hyper-parameters, it trains a model and computes an average
    validation error based on a TimeSeriesSplit
    """
    # pick hyper-parameters
    # hyperparams = {
    #     "metric": 'mae',
    #     "verbose": -1,
    #     "num_leaves": trial.suggest_int("num_leaves", 2, 256),
    #     "feature_fraction": trial.suggest_float("feature_fraction", 0.2, 1.0),
    #     "bagging_fraction": trial.suggest_float("bagging_fraction", 0.2, 1.0),
    #     "min_child_samples": trial.suggest_int("min_child_samples", 3, 100),   
    # }


    hyperparams = {
    "iterations": trial.suggest_int("iterations", 100, 1000),
    "learning_rate": trial.suggest_float("learning_rate", 1e-3, 1e-1, log=True),
    "depth": trial.suggest_int("depth", 4, 10),
    "l2_leaf_reg": trial.suggest_float("l2_leaf_reg", 1e-8, 100.0, log=True),
    "bootstrap_type": trial.suggest_categorical("bootstrap_type", ["Bayesian"]),
    "random_strength": trial.suggest_float("random_strength", 1e-8, 10.0, log=True),
    "bagging_temperature": trial.suggest_float("bagging_temperature", 0.0, 10.0),
    "od_type": trial.suggest_categorical("od_type", ["IncToDec", "Iter"]),
    "od_wait": trial.suggest_int("od_wait", 10, 50),
    "verbose": False
    }

       
    tss = TimeSeriesSplit(n_splits=2)
    scores = []
    for train_index, val_index in tss.split(X_train):

        # split data for training and validation
        X_train_, X_val_ = X_train.iloc[train_index, :], X_train.iloc[val_index,:]
        y_train_, y_val_ = y_train.iloc[train_index], y_train.iloc[val_index]
        
        # train the model
        pipeline = get_pipeline(**hyperparams)
        pipeline.fit(X_train_, y_train_)
        
        # evaluate the model
        y_pred = pipeline.predict(X_val_)
        mae = mean_absolute_error(y_val_, y_pred)

        scores.append(mae)
   
    # Return the mean score
    return np.array(scores).mean()

study = optuna.create_study(direction="minimize")
study.optimize(objective, n_trials=5)

best_params = study.best_trial.params
print(f'{best_params=}')

pipeline = get_pipeline(**best_params)
pipeline.fit(X_train, y_train)

predictions = pipeline.predict(X_test)
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

from plot import plot_one_sample

plot_one_sample(
    example_id=8,
    features=X_test,
    targets=y_test,
    predictions=pd.Series(predictions)
)
