==> notebooks/11_backfill_feature_store.ipynb <==
import os
from dotenv import load_dotenv
from paths import PARENT_DIR
from datetime import datetime
import pandas as pd
from functions import *
import hopsworks


# Preparing the environment and Constants

load_dotenv(PARENT_DIR / '.env')
PROJECT_NAME = 'nyc_taxi_pred'
HWORK_API_KEY = os.environ['HOPSWORK_API_KEY']

from_year = 2021
to_year = datetime.now().year

rides = pd.DataFrame()
for year in range(from_year, to_year+1):
    rides_one_year = load_raw_data(year)
    rides = pd.concat([rides, rides_one_year])




ts_data = transform_raw_data_into_ts_data(rides)

project = hopsworks.login(
    project=PROJECT_NAME,
    api_key_value=HWORK_API_KEY
    )

feature_store = project.get_feature_store()

FEATURE_GROUP_NAME = "time_series_hourly_feature_group"
FEATURE_GROUP_VERSION = 1

# Create feature group

feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description='time-series data at hourly frequency',
    primary_key=['pickup_location_id', 'pickup_hour'],
    event_time='pickup_hour',
)

feature_group.insert(ts_data,write_options={'wait_for_job':False})
